chore(core): clean clippy::all warnings in amlich-core and amlich-tui (amlich-081)
Resolve all cargo clippy --all-targets -- -W clippy::all warnings for
amlich-core and amlich-tui (lib + integration tests).

Categories addressed:
- doc_lazy_continuation / doc_overindented_list_items (~80): indent or
  re-indent test-function doc comments so continuation lines align with
  their list-item marker.
- assertions_on_constants (2): drop assert!(true, ...) informational
  stubs in khcbppt_taboos.rs and khcbppt_stars.rs (the eprintln report
  above already conveys the diagnostic).
- needless_range_loop (~18): convert for i in 0..N { arr[i] } loops to
  iter().enumerate() across almanac/fengshui, almanac/thai_tue,
  almanac/xung_hop, and two integration tests.
- unnecessary_to_owned (2): remove .to_string() on &str borrowed in
  HashSet::contains checks.
- sort_by_key (3): collapse descending sort_by closures into
  sort_by_key with std::cmp::Reverse on the count field.
- format_in_format_args (1): inline nested format! into a single
  format_args invocation in day_snapshot offering id construction.
- too_many_arguments (3): #[allow(clippy::too_many_arguments)] on
  direction_composite::assemble_cross_link,
  initiation_opening_evaluator::synthesize_primary_conclusion, and
  semantic_graph builders::interaction::add_ten_god_relation_node —
  internal helpers where a parameter struct would hurt readability.
- empty_line_after_doc_comments (1): drop the orphan duplicate doc
  comment that sat between get_days_to_nearest_tiet_khi and its tests.

Mechanical cargo clippy --fix also resolved:
- collapsible_if / items_after_test_module (graph_inspector.rs test
  module relocated to end of file).
- manual_range_contains / manual_is_multiple_of / redundant_clone /
  impl_can_derive / map_or etc. across core and tui.

After: cargo clippy -p amlich-core -p amlich-tui --all-targets -- -W clippy::all
reports zero warnings. cargo fmt -p amlich-core -p amlich-tui clean.
All existing tests still pass (848 amlich-core + 124 amlich-tui +
amlich-api). Whole workspace builds.

+++ crates/amlich-core/src/canchi.rs
@@ -72,7 +72,7 @@ pub fn get_month_canchi(lunar_month: i32, lunar_year: i32, is_leap_month: bool) 
     let first_month_can = first_month_can_table[year_can_index];
 
     // Calculate current month stem (offset from month 1)
-    let can_index = normalize_index(first_month_can as i32 + (lunar_month - 1), 10);
+    let can_index = normalize_index(first_month_can + (lunar_month - 1), 10);
 
     let mut result = CanChi::new(can_index, chi_index);
